U 12 use user id instead of hardcode

+++ backend/Controllers/RecipeController.cs
@@ -4,6 +4,7 @@ using FamilyMealPlanner.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
 using NLog;
+using System.Security.Claims;
 
 namespace FamilyMealPlanner.Controllers;
 
@@ -41,6 +42,9 @@ public class RecipeController(IWebScrappingService webScrappingService, IRecipeS
     [HttpPost("")]
     public async Task<IActionResult> Add(RecipeRequest recipe)
     {
+        if (!int.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out int userId))
+            return Unauthorized();
+
         try
         {
             int recipeId = await _recipeService.AddRecipe(recipe);
@@ -57,6 +61,9 @@ public class RecipeController(IWebScrappingService webScrappingService, IRecipeS
     [HttpGet("{recipeId}")]
     public async Task<IActionResult> GetById([FromRoute] int recipeId)
     {
+        if (!int.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out int userId))
+            return Unauthorized();
+
         try
         {
             Recipe recipe = await _recipeService.GetRecipeById(recipeId);
@@ -71,8 +78,11 @@ public class RecipeController(IWebScrappingService webScrappingService, IRecipeS
     }
 
     [HttpGet]
-    public async Task<IActionResult> GetByUserId([FromQuery] int userId)
+    public async Task<IActionResult> GetByUserId()
     {
+        if (!int.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out int userId))
+            return Unauthorized();
+
         try
         {
             List<Recipe> recipes = await _recipeService.GetRecipeByUserId(userId);
@@ -89,6 +99,9 @@ public class RecipeController(IWebScrappingService webScrappingService, IRecipeS
     [HttpPut("{recipeId}")]
     public async Task<IActionResult> Update(RecipeRequest recipeRequest, [FromRoute] int recipeId)
     {
+        if (!int.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out int userId))
+            return Unauthorized();
+
         try
         {
             await _recipeService.UpdateRecipe(recipeRequest, recipeId);
@@ -106,6 +119,9 @@ public class RecipeController(IWebScrappingService webScrappingService, IRecipeS
     [HttpDelete("{recipeId}")]
     public async Task<IActionResult> Delete([FromRoute] int recipeId)
     {
+        if (!int.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out int userId))
+            return Unauthorized();
+
         try
         {
             await _recipeService.Delete(recipeId);
